# openapi.yaml
openapi: 3.1.0
info:
  title: Image Redaction Service API
  version: 0.4.0
  description: |
    REST API for redacting sensitive regions in one or more images and converting
    each page of a PDF document to a JPEG image. The same capabilities are also
    exposed through MCP Streamable HTTP at /mcp.
servers:
  - url: http://localhost:3000
    description: Local development server
tags:
  - name: Health
    description: Service status endpoints.
  - name: Images
    description: Single-image and batch image redaction operations.
  - name: PDFs
    description: PDF rendering operations.
  - name: MCP
    description: MCP Streamable HTTP endpoint.
  - name: Documentation
    description: API description files.
paths:
  /health:
    get:
      tags: [Health]
      operationId: getHealth
      summary: Get service health, version, and capability metadata.
      responses:
        '200':
          description: Service is healthy.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
  /openapi.yaml:
    get:
      tags: [Documentation]
      operationId: getOpenApiYaml
      summary: Get the OpenAPI document for the REST service.
      responses:
        '200':
          description: OpenAPI YAML document.
          content:
            application/yaml:
              schema:
                type: string
  /v1/images/blur-sensitive-regions:
    post:
      tags: [Images]
      operationId: blurSensitiveRegions
      summary: Blur sensitive regions in one image.
      description: |
        Accepts an image as a data URL, raw base64, or a remote URL when remote
        image fetching is enabled. Region coordinates are normalized from 0 to 1.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BlurSensitiveRegionsRequest'
      responses:
        '200':
          description: Redacted image and processing metadata.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlurSensitiveRegionsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '408':
          $ref: '#/components/responses/RequestTimeout'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/images/blur-sensitive-regions/batch:
    post:
      tags: [Images]
      operationId: blurSensitiveRegionsBatch
      summary: Blur sensitive regions in multiple images concurrently.
      description: |
        Processes multiple image requests concurrently while preserving input order.
        Each image can have its own identifier, regions, output format, and options.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchBlurRequest'
      responses:
        '200':
          description: One redacted result per input image.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchBlurResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '408':
          $ref: '#/components/responses/RequestTimeout'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/pdfs/to-jpeg:
    post:
      tags: [PDFs]
      operationId: convertPdfToJpeg
      summary: Convert every PDF page to one JPEG image.
      description: |
        Accepts a PDF data URL, raw base64, or a remote URL when remote PDF fetching
        is enabled. The response contains one JPEG payload per page in page order.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PdfToJpegRequest'
      responses:
        '200':
          description: JPEG images generated from the PDF pages.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PdfToJpegResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '408':
          $ref: '#/components/responses/RequestTimeout'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /mcp:
    post:
      tags: [MCP]
      operationId: postMcp
      summary: MCP Streamable HTTP endpoint.
      description: |
        Stateless MCP endpoint exposing blur_sensitive_regions,
        blur_sensitive_regions_batch, and convert_pdf_to_jpeg.
      parameters:
        - name: Accept
          in: header
          required: false
          schema:
            type: string
            default: application/json, text/event-stream
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JsonRpcRequest'
      responses:
        '200':
          description: MCP JSON-RPC response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JsonRpcResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
    get:
      tags: [MCP]
      operationId: getMcp
      summary: Unsupported for this stateless MCP endpoint.
      responses:
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
    delete:
      tags: [MCP]
      operationId: deleteMcp
      summary: Unsupported for this stateless MCP endpoint.
      responses:
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
components:
  schemas:
    HealthResponse:
      type: object
      required: [status, version, capabilities, mcp]
      properties:
        status:
          type: string
          enum: [ok]
        version:
          type: string
          examples: [0.4.0]
        capabilities:
          type: object
          required: [singleImageRedaction, batchImageRedaction, pdfToJpeg]
          properties:
            singleImageRedaction:
              type: boolean
            batchImageRedaction:
              type: boolean
            pdfToJpeg:
              type: boolean
        mcp:
          type: object
          required: [transport, path]
          properties:
            transport:
              type: string
              enum: [streamable-http]
            path:
              type: string
              enum: [/mcp]
    FetchHeaders:
      type: object
      description: Only Authorization, Accept, and User-Agent headers are accepted.
      additionalProperties:
        type: string
    Region:
      type: object
      additionalProperties: false
      required: [x, y, width, height]
      properties:
        x:
          type: number
          minimum: 0
          maximum: 1
          description: Normalized left coordinate.
        y:
          type: number
          minimum: 0
          maximum: 1
          description: Normalized top coordinate.
        width:
          type: number
          exclusiveMinimum: 0
          maximum: 1
          description: Normalized region width.
        height:
          type: number
          exclusiveMinimum: 0
          maximum: 1
          description: Normalized region height.
        paddingPixels:
          type: integer
          minimum: 0
          maximum: 500
        blurRadius:
          type: number
          minimum: 0.3
          maximum: 100
    BlurOptions:
      type: object
      additionalProperties: false
      properties:
        blurRadius:
          type: number
          minimum: 0.3
          maximum: 100
          default: 14
        outputType:
          type: string
          enum: [image/jpeg, image/jpg, image/png, image/webp, image/avif]
          default: image/jpeg
        quality:
          type: integer
          minimum: 1
          maximum: 100
          default: 92
        returnDataUrl:
          type: boolean
          default: true
        inputMimeType:
          type: string
          default: image/jpeg
        regionPaddingPixels:
          type: integer
          minimum: 0
          maximum: 500
          default: 0
        fetchHeaders:
          $ref: '#/components/schemas/FetchHeaders'
    BlurSensitiveRegionsRequest:
      type: object
      additionalProperties: false
      required: [imageSource, regions]
      properties:
        imageSource:
          type: string
          description: Raw base64, data URL, or HTTP/HTTPS URL when enabled.
        regions:
          type: array
          minItems: 1
          maxItems: 100
          items:
            $ref: '#/components/schemas/Region'
        options:
          $ref: '#/components/schemas/BlurOptions'
    BatchBlurItem:
      type: object
      additionalProperties: false
      required: [imageSource, regions]
      properties:
        id:
          type: string
          description: Optional caller-defined identifier returned unchanged.
        imageSource:
          type: string
          description: Raw base64, data URL, or HTTP/HTTPS URL when enabled.
        regions:
          type: array
          minItems: 1
          maxItems: 100
          items:
            $ref: '#/components/schemas/Region'
        options:
          $ref: '#/components/schemas/BlurOptions'
    BatchBlurRequest:
      type: object
      additionalProperties: false
      required: [images]
      properties:
        images:
          type: array
          minItems: 1
          maxItems: 10
          items:
            $ref: '#/components/schemas/BatchBlurItem'
    BlurSensitiveRegionsResponse:
      type: object
      required: [image, mimeType, outputFormat, width, height, regionsProcessed]
      properties:
        id:
          type: string
        image:
          type: string
          description: Output image as a data URL or raw base64.
        mimeType:
          type: string
          enum: [image/jpeg, image/png, image/webp, image/avif]
        outputFormat:
          type: string
          enum: [jpeg, png, webp, avif]
        width:
          type: integer
          minimum: 1
        height:
          type: integer
          minimum: 1
        regionsProcessed:
          type: integer
          minimum: 0
    BatchBlurResponse:
      type: object
      required: [images, imagesProcessed, regionsProcessed]
      properties:
        images:
          type: array
          items:
            $ref: '#/components/schemas/BlurSensitiveRegionsResponse'
        imagesProcessed:
          type: integer
          minimum: 1
        regionsProcessed:
          type: integer
          minimum: 0
    PdfOptions:
      type: object
      additionalProperties: false
      properties:
        quality:
          type: integer
          minimum: 1
          maximum: 100
          default: 90
        scale:
          type: number
          minimum: 0.1
          maximum: 4
          default: 2
          description: Rendering scale applied to each page.
        returnDataUrl:
          type: boolean
          default: true
        fetchHeaders:
          $ref: '#/components/schemas/FetchHeaders'
    PdfToJpegRequest:
      type: object
      additionalProperties: false
      required: [pdfSource]
      properties:
        pdfSource:
          type: string
          description: Raw base64 PDF, PDF data URL, or HTTP/HTTPS URL when enabled.
        options:
          $ref: '#/components/schemas/PdfOptions'
    PdfPage:
      type: object
      required: [pageNumber, image, mimeType, width, height]
      properties:
        pageNumber:
          type: integer
          minimum: 1
        image:
          type: string
          description: JPEG page image as a data URL or raw base64.
        mimeType:
          type: string
          enum: [image/jpeg]
        width:
          type: integer
          minimum: 1
        height:
          type: integer
          minimum: 1
    PdfToJpegResponse:
      type: object
      required: [pages, totalPages, pagesConverted, mimeType]
      properties:
        pages:
          type: array
          items:
            $ref: '#/components/schemas/PdfPage'
        totalPages:
          type: integer
          minimum: 1
        pagesConverted:
          type: integer
          minimum: 1
        mimeType:
          type: string
          enum: [image/jpeg]
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
    JsonRpcRequest:
      type: object
      required: [jsonrpc, method]
      properties:
        jsonrpc:
          type: string
          enum: ['2.0']
        id:
          oneOf:
            - type: string
            - type: integer
            - type: 'null'
        method:
          type: string
        params:
          type: object
          additionalProperties: true
      additionalProperties: true
    JsonRpcResponse:
      type: object
      required: [jsonrpc]
      properties:
        jsonrpc:
          type: string
          enum: ['2.0']
        id:
          oneOf:
            - type: string
            - type: integer
            - type: 'null'
        result:
          type: object
          additionalProperties: true
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: integer
            message:
              type: string
      additionalProperties: true
  responses:
    BadRequest:
      description: Invalid request, image, or PDF input.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    RequestTimeout:
      description: Remote source fetch timed out.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    PayloadTooLarge:
      description: Input size, batch size, region count, or PDF page count exceeds configured limits.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    MethodNotAllowed:
      description: HTTP method is not supported for this endpoint.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/JsonRpcResponse'
    InternalServerError:
      description: Unexpected server-side error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'